- cosmetic
git-svn-id: http://svn.codergrid.de/svn/sf/bruteripper/trunk@239 7add9c9f-332d-0410-adcc-f7b1da7091d4

+++ test/test_bla.cpp
@@ -317,7 +317,7 @@ void writePPM(Frame  frame){
 
 
     // Open file
-  sprintf(szFilename, "frame%d.ppm", frame.pts);
+  sprintf(szFilename, "frame%d.ppm", (int)frame.pts);
   cout << "FileName:"<<szFilename<<endl;
   pFile=fopen(szFilename, "wb");
   if(pFile==NULL)
